Added the app challenges
Django will automatically look for apps templates only under 'templates'
folder, if the 'APP_DIRS' is set to True as in this case.

It is a convention that the folder MUST be named 'templates' AND the actual
template files should be put within a folder with the app name:
ie. templates > AppName > myTemplateFile.html